requirements.txt
python-dotenv
malaya
tensorflow
flask
